fix de l'erreur sur l'édition et la suppresion d'une catégorie

+++ app/Http/Controllers/CategoryController.php
@@ -81,7 +81,9 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
-        Storage::delete($category->cover_img);
+        if ($category->cover_img){
+            Storage::delete($category->cover_img);
+        }
         $category->delete();
         return redirect()->route('categories.index')->with("success","suppression de la catégorie reussie");
     }
